// sites/docs/src/lib/types/nav.ts
import type { Icons } from '$lib/components/docs/icons/index.js';

export type NavItem = {
    title: string;
    href?: string;
    disabled?: boolean;
    external?: boolean;
    icon?: keyof typeof Icons;
    label?: string;
};

export type SidebarNavItem = NavItem & {
    items: SidebarNavItem[];
};

export type NavItemWithChildren = NavItem & {
    items: NavItemWithChildren[];
};